+++ CHANGELOG.rst
@@ -252,6 +252,20 @@ Added
   `PR #903 <https://github.com/openego/eGon-data/pull/903>`_
 * Add low flex scenario 'eGon2035_lowflex'
   `#822 <https://github.com/openego/eGon-data/issues/822>`_
+* Fix depricated python-operator import
+  `#184 <https://github.com/openego/powerd-data/pull/184>`_
+* Add status2023 electrical neighbours
+  `#182 <https://github.com/openego/powerd-data/pull/182>`_
+* Fix transformer s_nom
+  `#168 <https://github.com/openego/powerd-data/pull/168>`_
+* Update osm for status2023
+  `#169 <https://github.com/openego/powerd-data/pull/169>`_
+* Use cached DemandRegio tables and scale status2023 cts & ind
+  `#185 <https://github.com/openego/powerd-data/pull/185>`_
+* Add status2023 scenario of demandregio households
+  `#186 <https://github.com/openego/powerd-data/pull/186>`_
+
+
 
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159